+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 #### 🚀 Updates
 
 - Added `type_relative_path` feature, that implements schematic types for the `relative-path` crate.
+- Added `type_url` feature, that implements schematic types for the `url` crate.
 
 ## 0.11.1