+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Fixes a bad release.
+
 ## 0.11.3
 
 #### 🚀 Updates